refactor(store): split global.ts grab-bag into focused submodules
Tier 1 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
global.ts (884 lines) was already section-comment-delineated into block
creation, wave-file fetching, counters, connection status, flash errors, tab
management, and misc utilities — extracted each into its own file
(window-identity/block-layout-actions/block-component-registry/wave-file/
conn-status/flash-notifications/tab-actions/dev-counters/misc-utils.ts), all
re-exported from global.ts alongside its existing backendStatus.ts/
block-atom-cache.ts re-exports, so none of the ~97 existing import sites
need to change. window-identity.ts is new — factored out because tab-actions
needed workspace/activeTabId which global.ts also needs directly; importing
them back from global.ts would have been circular, so both now import from
this shared base module instead.

global.ts: 884 -> 437 lines. tsc --noEmit clean; 164/164 tests across the 16
test files that import @/store/global pass unchanged.

<!-- agentmux:agent_id=agentx -->
